refactor: move detachProfileWithNextID
