+++ packages/mask/background/services/identity/profile/update.ts
@@ -1,4 +1,5 @@
-import type { ProfileIdentifier } from '@masknet/shared-base'
+import { NextIDAction, ProfileIdentifier } from '@masknet/shared-base'
+import { MaskMessages } from '../../../../shared/messages'
 import { storeAvatar } from '../../../database/avatar-cache/avatar'
 import {
     consistentPersonaDBWriteAccess,
@@ -6,6 +7,7 @@ import {
     deleteProfileDB,
     ProfileRecord,
 } from '../../../database/persona/db'
+import { NextIDProof } from '@masknet/web3-providers'
 
 export interface UpdateProfileInfo {
     nickname?: string | null
@@ -27,4 +29,18 @@ export async function updateProfileInfo(identifier: ProfileIdentifier, data: Upd
 export function mobile_removeProfile(id: ProfileIdentifier): Promise<void> {
     if (process.env.architecture !== 'app') throw new TypeError('This function is only available in app')
     return consistentPersonaDBWriteAccess((t) => deleteProfileDB(id, t))
+}
+
+export async function detachProfileWithNextID(
+    uuid: string,
+    personaPublicKey: string,
+    platform: string,
+    identity: string,
+    createdAt: string,
+    options?: { signature?: string },
+): Promise<void> {
+    await NextIDProof.bindProof(uuid, personaPublicKey, NextIDAction.Delete, platform, identity, createdAt, {
+        signature: options?.signature,
+    })
+    MaskMessages.events.ownProofChanged.sendToAll(undefined)
 }
