+++ packages/mask/src/extension/background-script/IdentityService.ts
@@ -20,7 +20,6 @@ import {
     ProfileInformation,
     PostIVIdentifier,
     RelationFavor,
-    NextIDAction,
 } from '@masknet/shared-base'
 import type { Persona, Profile } from '../../database/Persona/types'
 import {
@@ -46,7 +45,6 @@ import { getCurrentPersonaIdentifier } from './SettingsService'
 import { MaskMessages } from '../../utils'
 import { first, orderBy } from 'lodash-unified'
 import { recover_ECDH_256k1_KeyPair_ByMnemonicWord } from '../../utils/mnemonic-code'
-import { NextIDProof } from '@masknet/web3-providers'
 import { loginPersona } from '../../../background/services/identity/persona/update'
 
 assertEnvironment(Environment.ManifestBackground)
@@ -334,21 +332,3 @@ export async function queryPersonaByPrivateKey(privateKeyString: string) {
     return null
 }
 // #endregion
-
-export async function detachProfileWithNextID(
-    uuid: string,
-    personaPublicKey: string,
-    platform: string,
-    identity: string,
-    createdAt: string,
-    options?: {
-        walletSignature?: string
-        signature?: string
-        proofLocation?: string
-    },
-): Promise<void> {
-    await NextIDProof.bindProof(uuid, personaPublicKey, NextIDAction.Delete, platform, identity, createdAt, {
-        signature: options?.signature,
-    })
-    MaskMessages.events.ownProofChanged.sendToAll(undefined)
-}
